Report errors in resolver-loaded files
Previously, we would treat any error when requiring a file as a signal to move to the next potential file in the list.

This commit changes the behavior to only treat the “MODULE_NOT_FOUND” error as skippable—because the file doesn’t exist, not because there’s an error loading it—and re-throws other errors.