JSON API Support (#15)
Preliminary JSON API support

This commit adds preliminary support and architecture for quickly building JSON API 1.1-compliant endpoints.

In particular, it introduces to end users the concept of a `ResourceController` and a `Resource`. `ResourceControllers` are a subclass of controller that automatically try to parse incoming requests in terms of the JSON API semantics. The `Resource` class describes how to find, create, serialize and deserialize models for consumption via a JSON API.

Large chunks of this work and architecture are directly inspired by the JSONAPI::Resources gem for Rails: https://github.com/cerebris/jsonapi-resources. Big thanks to @lgebhardt and @dgeb!